+++ src/app/api/group/settlements/specyficExpense/route.ts
@@ -1,21 +1,22 @@
 import {NextRequest, NextResponse} from "next/server";
 import {fetchWithAuth} from "@/lib/api/fetch-with-auth";
+import {ExpenseCreate} from "@/lib/types/expense";
 
 const BASE_URL = process.env.BASE_URL;
-const DELETE_GET_PUT_RECOMMENDATIONS = process.env.DELETE_GET_PUT_RECOMMENDATIONS;
+const DELETE_GET_PUT_SETTLEMENTS = process.env.DELETE_GET_PUT_SETTLEMENTS;
 
 export async function GET(request: NextRequest) {
     try {
         const groupId = request.nextUrl.searchParams.get('groupId');
-        const recommendationId = request.nextUrl.searchParams.get('recommendationId');
-        if (!groupId || !recommendationId) {
+        const expenseId = request.nextUrl.searchParams.get('expenseId');
+        if (!groupId || !expenseId) {
             return NextResponse.json(
                 {success: false, message: 'Brak wymaganych parametrów'},
                 {status: 400}
             );
         }
-        const endpoint = DELETE_GET_PUT_RECOMMENDATIONS?.replace('{groupId}', groupId)
-            .replace('{recommendationId}', recommendationId);
+        const endpoint = DELETE_GET_PUT_SETTLEMENTS?.replace('{groupId}', groupId)
+            .replace('{expenseId}', expenseId);
         const cookieHeader = request.headers.get('cookie') ?? '';
         const response = await fetchWithAuth(`${BASE_URL}${endpoint}`, {
             method: 'GET',
@@ -29,7 +30,7 @@ export async function GET(request: NextRequest) {
         const data = await response.json();
         return NextResponse.json(data, {status: response.status});
     } catch (error) {
-        console.error('Recommendation retrieval API error:', error);
+        console.error('Expense retrieval API error:', error);
         return NextResponse.json(
             {success: false, message: 'Wystąpił błąd połączenia'},
             {status: 500}
@@ -39,9 +40,17 @@ export async function GET(request: NextRequest) {
 
 export async function PUT(request: NextRequest) {
     try {
-        const {title, category, imageURL, linkURL, content, groupId, recommendationId} = await request.json();
-        const endpoint = DELETE_GET_PUT_RECOMMENDATIONS?.replace('{groupId}', groupId)
-            .replace('{recommendationId}', recommendationId);
+        const groupId = request.nextUrl.searchParams.get('groupId');
+        const expenseId = request.nextUrl.searchParams.get('expenseId');
+        if (!groupId || !expenseId) {
+            return NextResponse.json(
+                {success: false, message: 'Brak wymaganych parametrów'},
+                {status: 400}
+            );
+        }
+        const expensePayLoad = await request.json() as ExpenseCreate;
+        const endpoint = DELETE_GET_PUT_SETTLEMENTS?.replace('{groupId}', groupId)
+            .replace('{expenseId}', expenseId);
         const cookieHeader = request.headers.get('cookie') ?? '';
         const response = await fetchWithAuth(`${BASE_URL}${endpoint}`, {
             method: 'PUT',
@@ -50,18 +59,14 @@ export async function PUT(request: NextRequest) {
                 'Content-Type': 'application/json',
             },
             body: JSON.stringify({
-                title,
-                content,
-                category,
-                imageUrl: imageURL,
-                linkURL
+                expensePayLoad
             }),
             credentials: 'include',
         });
         const data = await response.json();
         return NextResponse.json(data, {status: response.status});
     } catch (error) {
-        console.error('Recommendation update API error:', error);
+        console.error('Expense update API error:', error);
         return NextResponse.json(
             {success: false, message: 'Wystąpił błąd połączenia'},
             {status: 500}
@@ -71,9 +76,16 @@ export async function PUT(request: NextRequest) {
 
 export async function DELETE(request: NextRequest) {
     try {
-        const {groupId, recommendationId} = await request.json();
-        const endpoint = DELETE_GET_PUT_RECOMMENDATIONS?.replace('{groupId}', groupId)
-            .replace('{recommendationId}', recommendationId);
+        const groupId = request.nextUrl.searchParams.get('groupId');
+        const expenseId = request.nextUrl.searchParams.get('expenseId');
+        if (!groupId || !expenseId) {
+            return NextResponse.json(
+                {success: false, message: 'Brak wymaganych parametrów'},
+                {status: 400}
+            );
+        }
+        const endpoint = DELETE_GET_PUT_SETTLEMENTS?.replace('{groupId}', groupId)
+            .replace('{expenseId}', expenseId);
         const cookieHeader = request.headers.get('cookie') ?? '';
         const response = await fetchWithAuth(`${BASE_URL}${endpoint}`, {
             method: 'DELETE',
@@ -86,7 +98,7 @@ export async function DELETE(request: NextRequest) {
         const data = await response.json();
         return NextResponse.json(data, {status: response.status});
     } catch (error) {
-        console.error('Recommendation deletion API error:', error);
+        console.error('Expense deletion API error:', error);
         return NextResponse.json(
             {success: false, message: 'Wystąpił błąd połączenia'},
             {status: 500}
